+++ src/assets/css/cart.css
@@ -10,7 +10,7 @@
 .cart-item-qty {
   display: flex;
 }
-.art-item-remove-button {
+.cart-item-remove-button {
   display: flex;
   margin-left: 10px;
 }
